fix: dedup extracted requirements
