+++ duvet/src/text/find.rs
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+use super::whitespace;
 use core::{fmt, ops::Range};
 use triple_accel::levenshtein::levenshtein_search_simd_with_opts as text_search;
 
@@ -109,8 +110,8 @@ impl fmt::Debug for NormalizedSearch<'_> {
 
 impl<'a> NormalizedSearch<'a> {
     fn new(needle: &str, original_haystack: &'a str) -> Self {
-        let (needle, ()) = normalize_whitespace(needle);
-        let (haystack, offset_map) = normalize_whitespace(original_haystack);
+        let needle = whitespace::normalize(needle);
+        let (haystack, offset_map) = whitespace::normalize_mapped(original_haystack);
         Self {
             needle,
             haystack,
@@ -130,127 +131,6 @@ impl<'a> NormalizedSearch<'a> {
 
         Some(start..end)
     }
-}
-
-trait OffsetMap {
-    fn with_capacity(len: usize) -> Self;
-    fn push(&mut self, idx: usize);
-}
-
-impl OffsetMap for () {
-    #[inline]
-    fn with_capacity(_len: usize) -> Self {}
-
-    #[inline]
-    fn push(&mut self, _idx: usize) {}
-}
-
-impl OffsetMap for Vec<usize> {
-    #[inline]
-    fn with_capacity(len: usize) -> Self {
-        Vec::with_capacity(len + 1)
-    }
-
-    #[inline]
-    fn push(&mut self, idx: usize) {
-        self.push(idx);
-    }
-}
-
-fn normalize_whitespace<O: OffsetMap>(value: &str) -> (String, O) {
-    struct Mapper<O: OffsetMap> {
-        out: String,
-        offset_map: O,
-        buffer: Option<Buffer>,
-        last_end: usize,
-    }
-
-    impl<O: OffsetMap> Mapper<O> {
-        #[inline]
-        fn on_char(&mut self, idx: usize, c: char) {
-            if c.is_alphanumeric() {
-                self.flush();
-                self.push(idx, c);
-                return;
-            }
-
-            if c.is_whitespace() {
-                if self.buffer.is_none() && !self.out.is_empty() {
-                    self.buffer = Some(Buffer {
-                        start: idx,
-                        is_ws: true,
-                        c,
-                    });
-                }
-                return;
-            }
-
-            // punctuation
-            if let Some(buffer) = self.buffer.as_ref() {
-                if !buffer.is_ws {
-                    self.flush();
-                }
-            }
-
-            self.buffer = Some(Buffer {
-                start: idx,
-                is_ws: false,
-                c,
-            });
-        }
-
-        #[inline]
-        fn flush(&mut self) {
-            if let Some(buffer) = self.buffer.take() {
-                self.push(buffer.start, buffer.c);
-            }
-        }
-
-        #[inline]
-        fn push(&mut self, idx: usize, c: char) {
-            self.out.push(c);
-            let len = c.len_utf8();
-            for _ in 0..len {
-                self.offset_map.push(idx);
-            }
-            self.last_end = idx + len;
-        }
-
-        #[inline]
-        fn finish(mut self) -> (String, O) {
-            if let Some(buffer) = self.buffer.take() {
-                if !buffer.is_ws {
-                    self.push(buffer.start, buffer.c);
-                }
-            }
-            self.offset_map.push(self.last_end);
-            (self.out, self.offset_map)
-        }
-    }
-
-    struct Buffer {
-        start: usize,
-        is_ws: bool,
-        c: char,
-    }
-
-    let offset_map = O::with_capacity(value.len());
-    let out = String::with_capacity(value.len());
-
-    let mut mapper = Mapper {
-        offset_map,
-        out,
-        buffer: None,
-        last_end: 0,
-    };
-
-    for (idx, c) in value.char_indices() {
-        mapper.on_char(idx, c);
-    }
-
-    let (out, offset_map) = mapper.finish();
-
-    (out, offset_map)
 }
 
 #[cfg(test)]
@@ -300,7 +180,7 @@ mod tests {
     );
 
     fn normalize_whitespace(value: &str) -> (String, Vec<usize>) {
-        let (normalized, mapping) = super::normalize_whitespace::<Vec<usize>>(value);
+        let (normalized, mapping) = super::whitespace::normalize_mapped::<Vec<usize>>(value);
 
         dbg!(value, &normalized);
         let mut prev: Option<char> = None;
