+++ integration/extract-duplicate-requirement.toml
@@ -1,0 +1,31 @@
+source = { local = true }
+cmd = ["duvet report"]
+
+[[file]]
+path = ".duvet/config.toml"
+contents = """
+'$schema' = "https://awslabs.github.io/duvet/config/v0.4.0.json"
+
+[[source]]
+pattern = "src/my-code.rs"
+
+[[specification]]
+source = "my-spec.md"
+"""
+
+[[file]]
+path = "src/my-code.rs"
+contents = """
+//= my-spec.md#testing
+//# This MUST deduplicate the requirement.
+"""
+
+[[file]]
+path = "my-spec.md"
+contents = """
+# Testing
+
+This MUST deduplicate the requirement.
+
+This MUST deduplicate the requirement.
+"""
